+++ Nemesis-Mod/src/Hardware/GPIO/DigitalPin.cpp
@@ -3,7 +3,7 @@
 #include "DigitalPin.h"
 #include "Pin.h"
 
-DigitalPin::DigitalPin(uint32_t id) : Pin(id) {
+DigitalPin::DigitalPin(uint8_t id) : Pin(id) {
 }
 
 void DigitalPin::write(uint32_t value) {
